code cleanup & some nice features
variable names got renamed so its even more easy to understand.
weird parts of the code got removed and replaced with a better solution
option got added so upon adding a user to a group the groupids are stored to a list. each run the used ID will me removed from the list that is stored in memory
this avoids to send multiplie requests at the same time what then result in more groups users can be added to in a given time.